Add files via upload
Fixed recursion issue of $m$ needing to be larger than 2 in some use cases. Now an integer multiplier $m$ is found first and then $n$ begins to be used to refine the result +/- 1 from there. Additional protection put in for the case that no thickness is suitable to create at least one optimized repeat layer.